+++ src/components/internals/ToggleRow.js
@@ -1,9 +1,17 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-const ToggleRow = ({ record, recordApi, colSpan, style, className, children }) => (
+const ToggleRow = ({ 
+  record, 
+  recordApi, 
+  colSpan, 
+  toggleColumnHidden, 
+  style, 
+  className, 
+  children 
+}) => (
   <tr>
-    <td style={{ borderBottom: "1px solid transparent" }} colSpan={1} />
+    {!toggleColumnHidden && <td style={{ borderBottom: "1px solid transparent" }} colSpan={1} />}
     <td style={style || {}} className={`${className}`} colSpan={colSpan}>
       {children(record, recordApi)}
     </td>
@@ -14,6 +22,7 @@ ToggleRow.propTypes = {
   record: PropTypes.any,
   recordApi: PropTypes.any,
   colSpan: PropTypes.number,
+  toggleColumnHidden: PropTypes.bool,
   style: PropTypes.object,
   className: PropTypes.string
 };