+++ CHANGELOG.md
@@ -1,4 +1,16 @@
- ## 0.2.0 (April 24, 2018)
+## 0.3.0 (April 28, 2018)
+
+### QuickTable Component
+
+#### New Features
+
+* New `hideToggleColumn` prop - allows column with toggle icon to be hidden, useful when toggling the row via new Record Api method `toggleRecord()`
+
+### Record API
+
+* New `toggleRecord()` method - allows rows to be toggled via `recordApi` object passed to `Column` and `ToggleContent` components, useful when implementing row toggling via another element (ie button)
+
+## 0.2.0 (April 24, 2018)
 
 ### Column Component